// src/models/user.ts
export interface UserData {
  Full_Name: string;
  First_Name: string;
  TMC: string;
  Phone_No: string;
  Mobile_Phone_No: string;
  Personal_E_Mail: string;
  Location_Code: string;
  Gender: number;
  Designation: string;
  Department: string;
  Company_E_Mail: string;
  Date_of_Joining: string;
}

export const avatarCategory: any = {
  2: '/src/assets/images/men_avatar.png',
  1: '/src/assets/images/female_avatar.png',
};